Move codegen as a top-most package
