+++ pkg/codegen/pai/explain_render.go
