+++ pkg/codegen/tensorflow/template_explain.go
