+++ pkg/codegen/xgboost/docgen.go
