+++ pkg/sql/alisa_submitter.go
@@ -29,7 +29,7 @@ import (
 	"sqlflow.org/sqlflow/pkg/database"
 	"sqlflow.org/sqlflow/pkg/ir"
 	pb "sqlflow.org/sqlflow/pkg/proto"
-	"sqlflow.org/sqlflow/pkg/sql/codegen/pai"
+	"sqlflow.org/sqlflow/pkg/codegen/pai"
 )
 
 var resourceName = "job.tar.gz"
